update: corner case when folder have not a video

+++ include/ExploreFilesList.h
@@ -18,7 +18,7 @@ struct VideoFile
 	QString fileSize;
 	QString fps;
 
-	QString convertToByte(unsigned long long bit, bool isDecimal = false);
+	QString convertToByte(uint64_t bit, bool isDecimal = false);
 };
 
 class ExploreFilesList : public BaseDesignWindow
